[feature] fix file constructor
